added sum

+++ Model/src/External/Executable.java
@@ -15,6 +15,6 @@ public interface Executable {
     /**
      * Returns the value that from the execution of the command object
      */
-    public int returnValue();
+    public double returnValue();
 
 }
